ENH: Write an text file containing the list of the output images
Only 'important' output image file names are saved in this output file. This includes:
-All the low-rank and all the sparse images for the low-rank/sparse image decomposition.
-The atlas created by the Unbiased atlas creation.
-The average image as well as the low rank and sparse images computed during the final iteration for the low-rank atlas creation.
The file is named 'fileList.txt' and is written in the output directory. It is useful if one wants to quality control the outputs of the script that one has run. It is also useful when running the scripts from the pyLAR Slicer extension user interface. At the end of the process, this file can be used to know the name of the files to load in Slicer.