ENH: Accepts software paths and image paths including spaces
When running command lines through python, with subprocess.Popen,
passing arguments through a list instead of concatenating all
the arguments in a string. This allows to handle the case where
there are spaces in the software path as well as in the image
file names. There is no need to manually escape the space.