BUG: Cannot use Shell=True When using a list with subprocess.Popen()
subprocess.Popen() accepts a list of arguments or a string (command line)
as its first argument. Using a string causes problems with spaces in
filenames (problem escaping spaces). Using a list solves that issue,
but the option 'Shell=True' cannot be used. As a consequence, multiple
command lines cannot be concatenated into one command line and run
at the same time. This leads to not being able to run multiple registrations
in parallel when creating an unbiased atlas: We need to wait for the
registration result before the image can be transformed. The fact
that multple registrations cannot be run in parallel is not very limiting as
registrations can use many cores to be computed.